removed resolver controller and push the logic to the service. now modules get the id from a service and not a controller
