+++ src/bit-loader.js
@@ -10,7 +10,6 @@ var Transform  = require("./services/transform");
 var Dependency = require("./services/dependency");
 var Compile    = require("./services/compile");
 
-var Resolver   = require("./controllers/resolver");
 var Fetcher    = require("./controllers/fetcher");
 var Importer   = require("./controllers/importer");
 var Loader     = require("./controllers/loader");
@@ -65,7 +64,6 @@ function Bitloader(options) {
   // that build modules. Controllers use services, but services only use
   // services, not controllers.
   var controllers = {
-    resolver : new Resolver(this),
     fetcher  : new Fetcher(this),
     loader   : new Loader(this),
     importer : new Importer(this),
@@ -77,9 +75,8 @@ function Bitloader(options) {
 
   // Three methods as defined by:
   // https://whatwg.github.io/loader/#sec-properties-of-the-loader-prototype-object
-  this.import  = controllers.importer.import.bind(controllers.importer);
-  this.resolve = controllers.resolver.resolve.bind(controllers.resolver);
-  this.load    = controllers.loader.load.bind(controllers.loader);
+  this.import = controllers.importer.import.bind(controllers.importer);
+  this.load   = controllers.loader.load.bind(controllers.loader);
 
   this.fetch     = controllers.fetcher.fetch.bind(controllers.fetcher);
   this.important = controllers.importer.important.bind(controllers.importer);
@@ -122,7 +119,13 @@ Bitloader.prototype.import = function(/*names*/) {};
  * @returns {Promise} When resolved it returns a module meta object with
  *  the file path for the module.
  */
-Bitloader.prototype.resolve = function(/*name, referrer*/) {};
+Bitloader.prototype.resolve = function(name, referrer) {
+  return this.services.resolver
+    .resolve(new Module.Meta(name), referrer)
+    .then(function(moduleMeta) {
+      return moduleMeta.path;
+    });
+};
 
 
 /**
@@ -163,6 +166,18 @@ Bitloader.prototype.load = function(/*names, referrer*/) {};
 Bitloader.prototype.register = function(name, exports) {
   this.controllers.registry.register(name, exports);
   return this;
+};
+
+
+/**
+ * Method that determines if a module name is excluded from loading and processing.
+ *
+ * @param {string} name - Name of the module to test for exclusion.
+ *
+ * @returns {boolean}
+ */
+Bitloader.prototype.isExcluded = function(name) {
+  return this._exclude.indexOf(name) !== -1;
 };
